requirements.txt
typer[all]>=0.9.0
rich>=13.7.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
mcp>=0.9.0
aiohttp>=3.8.0